templateutils: Split package into multiple files
templateutils.go was getting a little large so some of the code has been
moved into some new files.

- defaults.go contains the function and help collections that represent the
  default set of functions you get if you specify a nil context.
- funcs.go contains the definition for all the new template functions provided
  by this package.
- usage.go contains the code that generates the description of the types exposed
  to the template program.
