+++ templateutils/usage.go
@@ -1,0 +1,102 @@
+//
+// Copyright (c) 2017 Intel Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package templateutils
+
+import (
+	"bytes"
+	"fmt"
+	"go/format"
+	"io"
+	"reflect"
+)
+
+func exportedFields(typ reflect.Type) bool {
+	for i := 0; i < typ.NumField(); i++ {
+		if typ.Field(i).PkgPath == "" {
+			return true
+		}
+	}
+
+	return false
+}
+
+func ignoreKind(kind reflect.Kind) bool {
+	return (kind == reflect.Chan) || (kind == reflect.Invalid)
+}
+
+func generateStruct(buf io.Writer, typ reflect.Type) {
+	for i := 0; i < typ.NumField(); i++ {
+		field := typ.Field(i)
+		if field.PkgPath != "" || ignoreKind(field.Type.Kind()) {
+			continue
+		}
+		fmt.Fprintf(buf, "%s ", field.Name)
+		tag := ""
+		if field.Tag != "" {
+			tag = fmt.Sprintf("`%s`", field.Tag)
+		}
+		generateUsage(buf, field.Type, tag)
+	}
+}
+
+func generateUsage(buf io.Writer, typ reflect.Type, tag string) {
+	kind := typ.Kind()
+	if ignoreKind(kind) {
+		return
+	}
+
+	switch kind {
+	case reflect.Struct:
+		if exportedFields(typ) {
+			fmt.Fprintf(buf, "struct {\n")
+			generateStruct(buf, typ)
+			fmt.Fprintf(buf, "}%s\n", tag)
+		} else if typ.Name() != "" {
+			fmt.Fprintf(buf, "%s%s\n", typ.String(), tag)
+		} else {
+			fmt.Fprintf(buf, "struct {\n}%s\n", tag)
+		}
+	case reflect.Slice:
+		fmt.Fprintf(buf, "[]")
+		generateUsage(buf, typ.Elem(), tag)
+	case reflect.Array:
+		fmt.Fprintf(buf, "[%d]", typ.Len())
+		generateUsage(buf, typ.Elem(), tag)
+	case reflect.Map:
+		fmt.Fprintf(buf, "map[%s]", typ.Key().String())
+		generateUsage(buf, typ.Elem(), tag)
+	default:
+		fmt.Fprintf(buf, "%s%s\n", typ.String(), tag)
+	}
+}
+
+func formatType(buf *bytes.Buffer, unformattedType []byte) {
+	const typePrefix = "type x "
+	source := bytes.NewBufferString(typePrefix)
+	_, _ = source.Write(unformattedType)
+	formattedType, err := format.Source(source.Bytes())
+	if err != nil {
+		panic(fmt.Errorf("formatType created invalid Go code: %v", err))
+	}
+	_, _ = buf.Write(formattedType[len(typePrefix):])
+}
+
+func generateIndentedUsage(buf *bytes.Buffer, i interface{}) {
+	var source bytes.Buffer
+	generateUsage(&source, reflect.TypeOf(i), "")
+	formatType(buf, source.Bytes())
+}
